pyproject.toml
[build-system]
requires = ["flit_core>=3.9"]
build-backend = "flit_core.buildapi"

[project]
name = "app"
version = "0.1.0"
description = "FastAPI-based Test Results API with PostgreSQL and Kubernetes Support"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
dependencies = [
    "click",
    "fastapi>=0.115.11",
    "loguru",
    "passlib[bcrypt]",
    "prometheus-fastapi-instrumentator",
    "psycopg2-binary",
    "pyjwt",
    "python-dotenv",
    "requests",
    "starlette>=0.46.1",
    "sqlmodel",
    "uvicorn[standard]",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
    "mypy",
    "pylint",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "radon",
]

[project.scripts]
api = "app.cli.main:cli"
test = "pytest:main"

[tool.flit.sdist]
include = ["app", "tests"]
exclude = ["__pycache__", "*.pyc", "*.pyo"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"